Added if/else in Ecomm/Address
If/Else statement deals with the case that there is a failure when
adding our credit card. This was added in the case the Litle doesn’t
handle our card and returns an error. Previously this was causing the
suite to crash so the if/else will handle this case gracefully,
allowing suite to continue.